Update QuitRate.py
Did experiments have distinct proportions of quit rate across qualification scores? To answer this I performed a Chi-squared test that evaluates if the two categories are independent. The null hypothesis is that the quit rate at a qualification score level is independent of the experiment that the person participated. If we can reject this hypothesis (p-value<0.05), then we will have strong evidence that quit rates across score levels are distinct (not independent) of the experiment that the person was part of. I could not reject the null-hypothesis (Chi2 stat=0.004, p-value=0.998, degrees of freedom=2). Hence, there is no evidence that a participant with a certain score level would have a distinct quit rate depending on the experiment that she participated.