Anova and Power test for Duration by Order of Tasks
The One Way Anova with Games-Howell correction for post hoc tests showed that
duration of first task was on aveage longer than the durations of the second
and third tasks (p-value<0.05). This is true for all Java methods. The only
exception was tasks 1 and 3 for Java method HIT07_33 (p-value=0.17).

The power test also showed the number of tasks needed to detect the differences
are in the order of magnitude of the experiment (below 1000 tasks).