Create ANOVA_TaskDurationByProfession_E2.R
ANOVA was significant. The following professions have distinct task durations (1st tasks in the assignment):
Hobbyist>Graduate_Student
Other>Graduate_Student
Professional_Developer<Hobbyist
Undergraduate_Student<Hobbyist
Professional_Developer<Other
Undergraduate_Student<Other
 YES, p_value = 3.2067647808294e-11
To detect these effects in 90% of the time, it would be necessary to have 1270 participants, which is in the order of magnitude of the study.