Update ANOVA_TaskDurationByProfession_E2.R
Are different professions faster to execute the 2nd and 3rd tasks in the assignments?
ANOVA test was statistical significant (p-value <0001). However, the power test showed that to detect the effect (distinct task durations across professions), it would be necessary 29 mil partipants. This 20 times larger than the number of participants in the experiment.