Update ANOVA_TaskDurationByOrder.R
Discussed posthoc results
Fixed analysis after fixing answer_index information for E2

The One Way Anova with Games-Howell correction for post hoc tests showed that
durations of first was on average longer than the durations of the second
or third tasks (p-value<0.05). This was true for ALL java methods.

The posthoc test results could not reject the null-hypothesis that second
and third task have same duration. Therefore, contrary to the first task effect on duration, we could not show the order effect on duration effect between second and third task. There was only one exception. Java method HIT03_6 presented statistical significant (p-value=0.03873688) differences between the duration of the second and third task. However, to detect this effect in 90% the time, we would need a number
of participants from 195 (HIT05_35) to 2013 (HIT01_8). This might not be realistic
when one is able to optimize the number of participants needed.